+++ tests/BlockCypher/Test/Api/AddressTest.php
@@ -203,7 +203,34 @@ class AddressTest extends \PHPUnit_Framework_TestCase
             ));
 
         /** @noinspection PhpParamsInspection */
-        $result = $obj->get("1DEP8i3QJCsomS4BSMY2RpU1upv62aGvhD", $mockApiContext, $mockBlockCypherRestCall);
+        $result = $obj->get("1DEP8i3QJCsomS4BSMY2RpU1upv62aGvhD", array(), $mockApiContext, $mockBlockCypherRestCall);
+        $this->assertNotNull($result);
+    }
+
+    /**
+     * @dataProvider mockProvider
+     * @param Address $obj
+     */
+    public function testGetWithParams($obj, /** @noinspection PhpDocSignatureInspection */
+                                      $mockApiContext)
+    {
+        $mockBlockCypherRestCall = $this->getMockBuilder('\BlockCypher\Transport\BlockCypherRestCall')
+            ->disableOriginalConstructor()
+            ->getMock();
+
+        $mockBlockCypherRestCall->expects($this->any())
+            ->method('execute')
+            ->will($this->returnValue(
+                AddressTest::getJson()
+            ));
+
+        $params = array(
+            'unspentOnly' => 'true',
+            'before' => 300000,
+        );
+
+        /** @noinspection PhpParamsInspection */
+        $result = $obj->get("1DEP8i3QJCsomS4BSMY2RpU1upv62aGvhD", $params, $mockApiContext, $mockBlockCypherRestCall);
         $this->assertNotNull($result);
     }
 
@@ -250,7 +277,7 @@ class AddressTest extends \PHPUnit_Framework_TestCase
 
         $addressList = Array(AddressTest::getObject()->getAddress());
 
-        $result = $obj->getMultiple($addressList, $mockApiContext, $mockBlockCypherRestCall);
+        $result = $obj->getMultiple($addressList, array(), $mockApiContext, $mockBlockCypherRestCall);
         $this->assertNotNull($result);
         $this->assertEquals($result[0], AddressTest::getObject());
     }
@@ -262,6 +289,33 @@ class AddressTest extends \PHPUnit_Framework_TestCase
     public static function getObject()
     {
         return new Address(self::getJson());
+    }
+
+    /**
+     * @dataProvider mockProvider
+     * @param Address $obj
+     */
+    public function testGetMultipleWithParams($obj, $mockApiContext)
+    {
+        $mockBlockCypherRestCall = $this->getMockBuilder('\BlockCypher\Transport\BlockCypherRestCall')
+            ->disableOriginalConstructor()
+            ->getMock();
+
+        $mockBlockCypherRestCall->expects($this->any())
+            ->method('execute')
+            ->will($this->returnValue(
+                '[' . AddressTest::getJson() . ']'
+            ));
+
+        $addressList = Array(AddressTest::getObject()->getAddress());
+        $params = array(
+            'unspentOnly' => 'true',
+            'before' => 300000,
+        );
+
+        $result = $obj->getMultiple($addressList, $params, $mockApiContext, $mockBlockCypherRestCall);
+        $this->assertNotNull($result);
+        $this->assertEquals($result[0], AddressTest::getObject());
     }
 
     public function mockProvider()